Fix remaining interpolated values
