+++ openaerostruct/structures/wingbox_group.py
@@ -2,6 +2,7 @@ import numpy as np
 import openmdao.api as om
 from openaerostruct.structures.section_properties_wingbox import SectionPropertiesWingbox
 from openaerostruct.structures.wingbox_geometry import WingboxGeometry
+from openaerostruct.utils.interpolation import get_normalized_span_coords
 
 
 class WingboxGroup(om.Group):
@@ -17,7 +18,7 @@ class WingboxGroup(om.Group):
         if "spar_thickness_cp" in surface.keys():
             n_cp = len(surface["spar_thickness_cp"])
             # Add bspline components for active bspline geometric variables.
-            x_interp = np.linspace(0.0, 1.0, int(ny - 1))
+            x_interp = get_normalized_span_coords(surface, mid_panel=True)
             comp = self.add_subsystem(
                 "spar_thickness_bsp",
                 om.SplineComp(
@@ -32,7 +33,7 @@ class WingboxGroup(om.Group):
         if "skin_thickness_cp" in surface.keys():
             n_cp = len(surface["skin_thickness_cp"])
             # Add bspline components for active bspline geometric variables.
-            x_interp = np.linspace(0.0, 1.0, int(ny - 1))
+            x_interp = get_normalized_span_coords(surface, mid_panel=True)
             comp = self.add_subsystem(
                 "skin_thickness_bsp",
                 om.SplineComp(
